First stab at config file template.
It just converts a salt dictionary to the json format used by ghost.
I'd like to keep this design if I can because it keeps the code
extremely simple. The intent is for the dictionary to be built by
taking formula defaults, updating with pillar default config, then
updating again with site-specific config.