Defaults changed to get the zip file from Ghost's site.
Unfortunately there is no url for "get the latest version, whatever
it is", and even if there were, I wouldn't have a matching hash
for it to put in defaults.yaml. So the default version has to be
hardcoded. That irritates me, because for now defaults.yaml will
have to be updated whenever a new version is released.